requirements.txt
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
redis==5.0.8
snowflake-connector-python==3.10.1
gunicorn==23.0.0
